unlockables for song_select_core part 1

+++ OpenTaiko/src/Lua/Classes/SongsManagement/LuaSongListSettings.cs
@@ -102,7 +102,7 @@ namespace OpenTaiko {
 
 			// Hide locked hidden songs
 			if (node.IsSong) {
-				if (node.HiddenIndex == DBSongUnlockables.EHiddenIndex.HIDDEN) return true;
+				if (node.HiddenIndex == (int)DBSongUnlockables.EHiddenIndex.HIDDEN) return true;
 			}
 
 			return false;
